requirements.txt
PyQt6>=6.5.0
pygame>=2.5.0
pynput>=1.7.0
google-auth>=2.23.0
google-auth-oauthlib>=0.9.0
google-api-python-client>=2.99.0
